Rename semaphore.h/c to lf_semaphore.h/c
